+++ MoreShipUpgrades/Plugin.cs
@@ -26,6 +26,7 @@ namespace MoreShipUpgrades
     [BepInDependency("com.github.WhiteSpike.CustomItemBehaviourLibrary", DependencyFlags.SoftDependency)]
     [BepInDependency(LethalLevelLoader.Plugin.ModGUID, DependencyFlags.SoftDependency)]
     [BepInDependency(Oxygen.OxygenBase.modGUID, DependencyFlags.SoftDependency)]
+    [BepInDependency(LCVR.Plugin.PLUGIN_GUID, DependencyFlags.SoftDependency)]
     public class Plugin : BaseUnityPlugin
     {
         internal static ManualLogSource mls;