+++ ChangeLog.md
@@ -1,4 +1,11 @@
 <details>
+<summary> 3.10.1 - 2024-08-29 </summary>
+
+- Possibly fixed issue with Back Muscles when using Lethal Company Virtual Reality mod.
+
+</details>
+
+<details>
 <summary> 3.10.0 - 2024-08-28 </summary>
 
 - Implemented "Oxygen Canister" tier upgrade which reduces the oxygen consumption rate of the player.